# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MoobilPay API",
        "description": "API complète de paiement en ligne avec Stripe et PayPal. Permet de gérer les paiements, méthodes de paiement et webhooks de manière sécurisée.",
        "contact": {
            "name": "Support MoobilPay",
            "email": "tchindavaldoblair@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement local"
        },
        {
            "url": "https://api.moobilpay.com",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/api/payment-methods/{id}": {
            "get": {
                "tags": [
                    "Payment Methods"
                ],
                "summary": "Détails d'une méthode de paiement",
                "operationId": "fb8ba779f90c67c04d432777f690aa20",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la méthode de paiement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la méthode de paiement"
                    },
                    "404": {
                        "description": "Méthode de paiement non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Payment Methods"
                ],
                "summary": "Modifier une méthode de paiement",
                "operationId": "477b007621815d6156c36639f5951fbd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "is_default": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Méthode de paiement modifiée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Payment Methods"
                ],
                "summary": "Supprimer une méthode de paiement",
                "operationId": "7b4e5ac6586e8decceec2d7a0f9aa1c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Méthode de paiement supprimée"
                    },
                    "404": {
                        "description": "Méthode de paiement non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payment-methods/{id}/set-default": {
            "post": {
                "tags": [
                    "Payment Methods"
                ],
                "summary": "Définir comme méthode de paiement par défaut",
                "operationId": "5b067a2cfdd7b2320896ba8bcdc89b74",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Méthode définie par défaut"
                    },
                    "404": {
                        "description": "Méthode de paiement non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payments/stripe/frontend/intent": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Créer un PaymentIntent (flux frontend)",
                "operationId": "7aa4ef289c29a8b906244b4dfa661522",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "example": 29.99
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "EUR"
                                    },
                                    "metadata": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Client secret retourné",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "payment_intent_id": {
                                            "type": "string"
                                        },
                                        "client_secret": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payments/stripe/frontend/confirm": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Confirmer un PaymentIntent (flux frontend)",
                "operationId": "aab5cdef66c648359e879eff900d4d2e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "payment_intent_id"
                                ],
                                "properties": {
                                    "payment_intent_id": {
                                        "type": "string",
                                        "example": "pi_123"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "pm_123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Statut du PaymentIntent"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification via Laravel Sanctum. Utilisez votre token d'API.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}